refactor(tunnel): telegram webhook registration moves into the channel's domain
tunnel.ts (generic cloudflared orchestration) carried the telegram-specific setWebhook logic and a
reverse dependency on the channel's transport internals. The tunnel keeps only the per-channel
DISPATCH ("this workspace has a telegram channel → register it"); what registering means lives in
channels/telegram/register-webhook.ts, beside the pipeline it rides. Behavior unchanged.
